refactor(core): split user concerns
